refactor(cli): regroup public commands and aliases
