+++ packages/core/cli/README.zh-CN.md
@@ -129,38 +129,30 @@ nb install --env app1 --resume
 | 命令 | 说明 |
 | --- | --- |
 | `nb init` | 初始化 NocoBase，并连接为 coding agent 可使用的 CLI env。 |
-| `nb install` | `nb init` 内部会使用的高级命令，用于安装本地 NocoBase 应用并保存 env 配置。通常建议直接使用 `nb init`。 |
-| `nb download` | `nb init` 或 `nb upgrade` 会使用的高级命令，用于从 Docker、npm 或 Git 获取 NocoBase。通常很少直接使用。 |
-| `nb start` | 启动选中的本地应用或 Docker 容器。 |
-| `nb stop` | 停止选中的本地应用或 Docker 容器。 |
-| `nb restart` | 先停止，再启动选中的本地应用或 Docker 容器。 |
-| `nb dev` | 为 npm/Git 源码 env 启动开发模式。 |
-| `nb logs` | 查看 npm/Git 或 Docker env 的应用日志。 |
-| `nb ps` | 查看已配置 env 的运行状态。 |
+| `nb app` | 管理应用运行态：启动、停止、重启、日志、状态、清理和升级。 |
+| `nb source` | 管理本地源码工程：下载、开发、构建和测试。 |
 | `nb db` | 查看或管理本地 env 的内置数据库运行状态。 |
-| `nb upgrade` | 更新源码/镜像并重启选中的应用。 |
-| `nb down` | 停止并移除某个 env 的本地运行容器。 |
 | `nb env` | 管理已保存的 CLI env 连接。 |
 | `nb api` | 通过 CLI 调用 NocoBase API 资源。 |
-| `nb pm` | 管理选中 NocoBase env 的插件。 |
+| `nb plugin` | 管理选中 NocoBase env 的插件。 |
 
 推荐在应用和运行时相关命令里显式使用 `--env`；`-e` 是它的简写：
 
 ```bash
-nb start --env app1
-nb restart --env app1
-nb logs --env app1
-nb ps --env app1
+nb app start --env app1
+nb app restart --env app1
+nb app logs --env app1
+nb app ps --env app1
 nb db ps --env app1
 ```
 
 等价的简写示例：
 
 ```bash
-nb start -e app1
-nb restart -e app1
-nb logs -e app1
-nb upgrade -e app1
+nb app start -e app1
+nb app restart -e app1
+nb app logs -e app1
+nb app upgrade -e app1
 nb db start -e app1
 ```
 
@@ -172,18 +164,18 @@ Docker env 会通过已保存的 Docker 容器和镜像进行管理：
 
 ```bash
 nb init --env app1 --yes --source docker --version alpha
-nb start --env app1
-nb restart --env app1
-nb logs --env app1
-nb stop --env app1
+nb app start --env app1
+nb app restart --env app1
+nb app logs --env app1
+nb app stop --env app1
 ```
 
 Docker 下载支持指定平台：
 
 ```bash
-nb download --source docker --version alpha --docker-platform auto
-nb download --source docker --version alpha --docker-platform linux/amd64
-nb download --source docker --version alpha --docker-platform linux/arm64
+nb source download --source docker --version alpha --docker-platform auto
+nb source download --source docker --version alpha --docker-platform linux/amd64
+nb source download --source docker --version alpha --docker-platform linux/arm64
 ```
 
 ### npm 和 Git
@@ -192,10 +184,10 @@ npm 和 Git env 会使用本地源码目录，并支持开发模式：
 
 ```bash
 nb init --env app1 --yes --source git --version alpha
-nb dev --env app1
+nb source dev --env app1
 ```
 
-`nb dev` 只支持 npm/Git 源码 env。Docker env 可以通过 `nb logs` 查看日志，remote env 只支持 API 和 env 相关操作。
+`nb source dev` 只支持 npm/Git 源码 env。Docker env 可以通过 `nb app logs` 查看日志，remote env 只支持 API 和 env 相关操作。
 
 ### 已有 NocoBase 应用
 
@@ -212,13 +204,13 @@ nb env add app1 --api-base-url http://localhost:13000/api
 升级会更新已保存的源码或镜像，然后重启应用：
 
 ```bash
-nb upgrade --env app1
+nb app upgrade --env app1
 ```
 
 如果只想使用当前已保存的本地源码或 Docker 镜像重启应用，可以使用 `--skip-code-update` 或 `-s`：
 
 ```bash
-nb upgrade --env app1 -s
+nb app upgrade --env app1 -s
 ```
 
 ## 数据库命令
