+++ packages/core/cli/src/commands/examples/prompts-stages.ts
@@ -12,13 +12,13 @@ import {
   type PromptInitialValues,
   type PromptsCatalog,
   runPromptCatalog,
-} from '../lib/prompt-catalog.ts';
+} from '../../lib/prompt-catalog.ts';
 import {
   applyCliLocale,
   CLI_LOCALE_FLAG_DESCRIPTION,
   CLI_LOCALE_FLAG_OPTIONS,
-} from '../lib/cli-locale.ts';
-import { type RunPromptCatalogWebUIStage, runPromptCatalogWebUI } from '../lib/prompt-web-ui.ts';
+} from '../../lib/cli-locale.ts';
+import { type RunPromptCatalogWebUIStage, runPromptCatalogWebUI } from '../../lib/prompt-web-ui.ts';
 import PromptsTest from './prompts-test.ts';
 
 function buildWebUiStagesFromTestPrompts(
